Add allong.es to functional programming category
Url to the package: https://www.npmjs.org/package/allong.es

allong.es is a library done by the guy who wrote one of the best books about functional programming in JS: JavaScript Allonge (https://leanpub.com/javascript-allonge) - the lib is complementary to the book (it's the implementation of what the book tells about).

From the lib's readme:
```
The allong.es library is a collection of functions designed to facilitate writing JavaScript and/or CoffeeScript with functions as first-class values. The emphasis in allong.es is on composing and decomposing functions using combinators and decorators. allong.es is designed to complement libraries like Underscore, not compete with them.
```