Added some Bevry projects
Sorry for them all being bundled, happy to modify the pull request to make it easier. I was a bit taken aback that none of the Bevry projects where on here (however our training was, so that was a start). Not sure if it's because we're an Australian company and just don't have a good reach to the outside world, or what's the go.