Adds a link to [ulog](https://github.com/download/ulog).
Ulog is a universal log library that combines features from the popular `debug` utility with log levels, creating a more generally applicable logger. It also adds better browser support, allowing for the log level to be changed via the query string in much the same way as we do with environment variables in Node JS.

When I started developing in Javascript I created a small logging shim to wrap the console for situations where it was not available. When I started developing in Node, I tried a bunch of logging frameworks, but none satisfied my needs. So I modified my own library based on some new ideas I had seen and created [picolog](https://github.com/download/picolog). I have been using that for about a year. In this time I discovered some things that lead to ulog:

1) I found that it's natural for the default log level to be INFO on Node, but WARN on the browser
2) When there are lots of modules doing logging, you end up wanting to see only the messages from some of them
2) I discovered the popularity of debug. It's named modules and filtering idea is brilliant, but I did not like the fact that this was purely aimed at debugging. I wanted something more general purpose.
3) I stumbled on the author of `ulog`, a project started but never completed. He agreed to let me use the name

So I created ulog. I have been using it for a month now. I have some ideas to make it even better (mostly I want to be able to use it as an alias for debuug so I can force packages using debug to use ulog instead), but I feel it's already full-featured so I'll do that when the need arises.

The package has a test suite and CI integration. I tested it's predecessor on platforms such as Nashorn and with loaders such as require js to make sure it is truly universal and I've been personally using and tweaking it for a very long time.