Add git-guppy commandline app
git-guppy is a tool that works in conjunction with your gulp setup to automatically configure git hooks that trigger gulp tasks. 

It's a mature/stable project that's being used by a lot of projects in production-ready situations.

It's not quite a commandline util, not quite a build tool, and it has standalone terminal commands (optional install). So not easy to categorize, but it seems that commandline app is the closest fit.