Add NW.js
It's been around 5 years longer than Electron, it's updated faster and more often (within 24 hours of any Node or Chromium release), has new features added faster, produces smaller distribution sizes for twice as many operating systems, and is much easier to use. It's downside is that it doesn't have GitHub advertising it, so it would be nice for it not to be exclude from this list.