Added hashnode nodejs
Hashnode nodejs is a quick drip nodejs community for those who want a faster post rate than the Reddit, and a less formal question asking place than the Stack Overflow. A community of almost 10k, very active and helpful when wanting to see the newest of Node or ask a quick question to get a quick response.