Add Expressa to "Web frameworks"
Add expressa which is a middleware that lets you make APIs from JSON schemas. It provides an editor that lets you edit your data as well as modify your schema and manage permissions for different operations. This is not just another web framework in that 1) it's easy to integrate into other applications since it's just a middleware 2) it's declarative as your schemas and permissions are defined in JSON files 3) It comes with an admin editor.