Add Video: CI/CD Pipeline with Node.js and GitHub Actions
Hi,

I'm the CTO and Co-Founder of datree.io and AWS Community Hero. This is a hands-on Webinar with live coding that I'm presenting about how to build a CI/CD pipeline with GitHub Actions, Node.js and AWS ECS Fargate. We have hundreds of RSVPs and I genuinely think that it's valuable to this community.

https://dtr.ee/2DURIVu

Let me know what you think!

Best regards,
Shimon Tolts

**By submitting this pull request, I promise I have read the [contribution guidelines](https://github.com/sindresorhus/awesome-nodejs/blob/master/contributing.md) twice and ensured my submission follows it. I realize not doing so wastes the maintainers' time that they could have spent making the world better. 🖖**

⬆⬆⬆⬆⬆⬆⬆⬆⬆⬆