Add jsep tiny expression parser package
This package is a lot more lightweight than a full AST parser, so it is easier for users to get started with.

I've used it in a project and it was a breeze to setup. The documentation is also really cool, since its just an annotated version of the source code,and still it is very clear and descriptive. 

https://github.com/soney/jsep